Move App component
