+++ src/components/App/test.tsx
@@ -6,7 +6,7 @@ import { render, screen } from '@testing-library/react';
 
 import '@testing-library/jest-dom';
 
-import App from '../App';
+import App from '.';
 
 describe('<App />', () => {
   it('renders without errors', () => {
