+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 
 import { ChakraProvider, ColorModeScript } from '@chakra-ui/react';
 
-import App from './App';
+import App from './components/App';
 import theme from './theme';
 
 ReactDOM.render(
